Enhanced Zion Tech Group website with comprehensive AI services, improved navigation, and mobile responsiveness
- Added 15+ innovative AI micro SaaS, IT services, and advanced AI solutions
- Implemented real market pricing with competitive rates and ROI metrics
- Created detailed service pages with features, benefits, and contact information
- Enhanced navigation with organized service categories and pricing
- Improved mobile responsiveness across all screen sizes
- Added comprehensive services data structure with detailed information
- Updated header and footer with better organization and contact details
- Ensured all content fits properly on mobile, tablet, and desktop
- Successfully tested build process and verified functionality
